+++ src/main/scala/longevity/subdomain/embeddable/DerivedType.scala
@@ -1,11 +1,12 @@
 package longevity.subdomain.embeddable
 
+import emblem.TypeKey
+
 /** one of the derived types in a family of domain entity types. mix this in to
  * your [[EType]] when it represents a concrete subtype of a [[PolyType]].
  */
-trait DerivedType[P <: Embeddable, Poly >: P <: Embeddable] extends EType[P] {
+abstract class DerivedType[P <: Embeddable : TypeKey, Poly >: P <: Embeddable : TypeKey] extends EType[P] {
 
-  /** the poly type that this type is derived from */
-  val polyType: PolyType[Poly]
+  private[longevity] val polyTypeKey: TypeKey[Poly] = implicitly[TypeKey[Poly]]
 
 }