+++ src/test/scala/longevity/integration/subdomain/controlledVocab/ControlledVocab.scala
@@ -12,18 +12,12 @@ object ControlledVocab extends EType[ControlledVocab] with PolyType[ControlledVo
 // empty case class version:
 case class VocabTerm1() extends ControlledVocab
 
-object VocabTerm1 extends DerivedType[VocabTerm1, ControlledVocab] {
-  val polyType = ControlledVocab
-}
+object VocabTerm1 extends DerivedType[VocabTerm1, ControlledVocab]
 
 // case object version:
 case object VocabTerm2 extends ControlledVocab
 
-object VocabTerm2Type extends DerivedType[VocabTerm2.type, ControlledVocab] {
-  val polyType = ControlledVocab
-}
+object VocabTerm2Type extends DerivedType[VocabTerm2.type, ControlledVocab]
 
-// illegal cyclic reference:
-// case object VocabTerm2 extends ControlledVocab with DerivedType[VocabTerm2.type, ControlledVocab] {
-//   val polyType = ControlledVocab
-// }
+// this doesnt work cuz illegal cyclic reference:
+// case object VocabTerm2 extends DerivedType[VocabTerm2.type, ControlledVocab] with ControlledVocab