+++ src/test/scala/longevity/integration/subdomain/derivedEntities/FirstDerivedEntity.scala
@@ -7,7 +7,4 @@ case class FirstDerivedEntity(
   first: String)
 extends PolyEntity
 
-object FirstDerivedEntity extends DerivedType[FirstDerivedEntity, PolyEntity] {
-  val polyType = PolyEntity
-}
-
+object FirstDerivedEntity extends DerivedType[FirstDerivedEntity, PolyEntity]